backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.3.0
python-dotenv==1.0.0
rapidfuzz==3.5.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.32.3
noisereduce==3.0.0
numpy==1.24.3
soundfile==0.13.1
boto3==1.34.0
scikit-learn==1.3.0
seaborn==0.12.2
matplotlib==3.7.2
setuptools>=65.0.0
wheel
sentence-transformers==2.2.2
transformers==4.30.2
huggingface-hub==0.13.4
torch==2.0.1
scikit-learn
numpy
twilio==8.10.0
